Adding a function for checking insertions
Added a function that checks if the user inserted any address, date, or hour in their edit boxes. If the user didn't made any necessary insertions, when he will press on the confirmation button it will throw a Toast and drag the screen to the right edit text where he didn't inserted the necessary information.